requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core==1.26.3
azure-functions==1.13.2
azure-storage-blob==12.14.1
requests==2.28.2
requests-oauthlib==1.3.1
